decompose contacts controller
