+++ controllers/contacts/add.js
@@ -1,0 +1,12 @@
+const { ctrlWrapper } = require("../../helpers");
+const { Contact } = require("../../models/contact");
+
+const add = async (req, res, next) => {
+  const result = await Contact.create(req.body);
+
+  res.status(201).json(result);
+};
+
+module.exports = {
+  add: ctrlWrapper(add),
+};
