+++ controllers/contacts/getAll.js
@@ -1,0 +1,12 @@
+const { ctrlWrapper } = require("../../helpers");
+const { Contact } = require("../../models/contact");
+
+const getAll = async (req, res, next) => {
+  const result = await Contact.find({}, "-createdAt -updatedAt");
+
+  res.json(result);
+};
+
+module.exports = {
+  getAll: ctrlWrapper(getAll),
+};
