+++ controllers/contacts/updateStatusContact.js
@@ -1,0 +1,17 @@
+const { requestError, ctrlWrapper } = require("../../helpers");
+const { Contact } = require("../../models/contact");
+
+const updateStatusContact = async (req, res, next) => {
+  const { contactId } = req.params;
+  const result = await Contact.findByIdAndUpdate(contactId, req.body, {
+    new: true,
+  });
+
+  if (!result) throw requestError(404);
+
+  res.json(result);
+};
+
+module.exports = {
+  updateStatusContact: ctrlWrapper(updateStatusContact),
+};
